requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
google-genai>=0.8
pydantic-settings>=2.6
python-dotenv>=1.0
